+++ src/tech/crcstraw.cc
@@ -91,4 +91,4 @@ int CRCStraw::Get(void* source, int slen) {
  *                                                                                             *
  * HISTORY: * 07/03/1996 JLB : Created. *
  *=============================================================================================*/
-long CRCStraw::Result() const { return crc_.Value(); }
+uint32_t CRCStraw::Result() const { return crc_.Value(); }
